requirements.txt
selenium
webdriver-manager
python-dotenv
pandas
PyMuPDF
openai
requests
pymupdf
colorama